Reorganizing some of the code
